Subject: Create Main class for XML unmarshalling
Body:
This commit introduces the Main class in the com.solvd package. The Main class contains the main method responsible for unmarshalling XML data into Java objects using JAXB (Java Architecture for XML Binding).

The main method instantiates a File object pointing to the XML file located at "src/main/resources/xml/xml.xml". Within a try-catch block, it creates a JAXBContext for the Laptop class and initializes an Unmarshaller. It then unmarshals the XML file into a Laptop object using the unmarshal() method.

If successful, the Laptop object is printed using the overridden toString() method to display the laptop information. In case of a JAXBException, a RuntimeException is thrown.

This class serves as an entry point to demonstrate XML unmarshalling functionality using JAXB.