Refactor 'Ram' class with Builder Pattern and Listener Pattern
This commit refactors the 'Ram' class by introducing the Builder Pattern to provide a more flexible and readable way of creating instances. Additionally, a private constructor and a static inner 'RamBuilder' class are added. The 'RamListener' interface is introduced for potential listeners interested in events related to 'Ram' creation.

Changed fields:
- Changed 'Computer_computer_id' to 'computer_computer_id'
- Introduced a private constructor for internal use
- Added a static inner 'RamBuilder' class for building instances
- Implemented a private constructor that takes a 'RamBuilder' for building instances

Tasks satisfied:
- Builder
- Listener